+++ tests/test_public_surface.py
@@ -9,13 +9,13 @@ from country_registry import get_country_config, list_exposed_countries, list_ex
 
 class PublicSurfaceContractTests(unittest.TestCase):
     def test_single_country_registry(self):
-        countries = list_exposed_countries(["netherlands"], allow_internal=True)
+        countries = list_exposed_countries(["netherlands"], allow_internal=False)
         self.assertEqual(len(countries), 1)
         self.assertEqual(countries[0].country_id, "netherlands")
 
     def test_public_visibility_flag(self):
         countries = list_exposed_public_countries()
-        self.assertEqual(len(countries), 0)
+        self.assertEqual(len(countries), 1)
 
     def test_country_config_matches_public_surface(self):
         config = get_country_config("netherlands")